requirements.txt
Flask==3.0.3
python-bring-api==3.0.0
python-dotenv==1.0.1